v2020.d.1 | First commit
----

Version: 2020.d.1
Date: April 24, 2020	23:00

----

This is the first commit of the project. The bootstrap program, lib and
main module are created. The application's GUI layout is properly set.

Basic saving and loading of file is implemented, but the FileDialog requires
extra work to make it work properly.

Light/Dark modes are planned to be implemented.

----

Readme and gitignores are updated.