v2020.d.2 | Completed File Dialog
----

Version: 2020.d.2
Date: April 25, 2020	20:52

----

The file dialog is properly implemented and is functioning. The experimental
version of the file dialog is used rather than the original one because it
supports different modes (open/save).

----

The text area is now placed inside a scroll view to enable scrolling.

----

A new preference window is added to provide app settings modifications.
When the "Preference" button is clicked, this window will pop up as a
new window. Everytime it will be created as a new object.