v2020.d.2 | Major Completion of Project
----

Version: 2020.d.2
Date: April 27, 2020	18:17

----

The projected is mainly completed, with all planned features being functional.

----

Save checking is implemented to check whether the current file is saved.
If the user closes the application and the current file is not yet saved,
a warning window will show up asking the user for confirmation.

----

The application is added with the preferences feature. Users can now set
the desired theme and display font size of the application via the preference
window. The preference data is saved to a embed file of the application and is
retrieved everytime when the application starts.