v2020.e.3 | Fixed Preferences Window Cancel Button Not Working
----

Version: 2020.e.3
Date: May 12, 2020   22:07   GMT+8

----

The Cancel button failed to work because the root preference values are set
directly set when options are selected. This issue is fixed by adding another
set of temporary preference values to the preferences window. When the window
starts, the values are copied from the root to it. The values are only applied
back to the root when the OK button is pressed, otherwise the original values
remain unchanged.